exception handle v1.1

+++ src/main/java/org/example/service/util/PackageServiceIMPL.java
@@ -3,6 +3,7 @@ package org.example.service.util;
 import lombok.RequiredArgsConstructor;
 import org.example.dto.PackageDTO;
 import org.example.entity.PackageEntity;
+import org.example.exception.NotFoundException;
 import org.example.repo.MainTravelServiceRepo;
 import org.example.repo.PackageRepo;
 import org.example.service.PackageService;
@@ -29,22 +30,36 @@ public class PackageServiceIMPL implements PackageService {
 
     @Override
     public PackageDTO getSelectPackage(String package_id) {
-        return converter.toPackageDto(packageRepo.findById(package_id).get());
+        Optional<PackageEntity> byId = packageRepo.findById(package_id);
+        if (!byId.isPresent()){
+            throw new NotFoundException("Package ID Not Found :"+package_id);
+        }
+        return converter.toPackageDto(byId.get());
     }
 
     @Override
-    public void updatePackage(PackageDTO packageDTO) {
-        Optional<PackageEntity> byId = packageRepo.findById(packageDTO.getPackage_id());
+    public void updatePackage(String package_id,PackageDTO packageDTO) {
+        Optional<PackageEntity> byId = packageRepo.findById(package_id);
         if (!byId.isPresent()){
-            byId.get().setPackage_cetagory(packageDTO.getPackage_cetagory());
-            byId.get().setPackage_hotel_cetagory(packageDTO.getPackage_hotel_cetagory());
-            byId.get().setPackage_hotel_cetagory(packageDTO.getPackage_hotel_cetagory());
+            throw new NotFoundException("Package ID Not Found :"+package_id);
+//            byId.get().setPackage_cetagory(packageDTO.getPackage_cetagory());
+//            byId.get().setPackage_hotel_cetagory(packageDTO.getPackage_hotel_cetagory());
+//            byId.get().setPackage_hotel_cetagory(packageDTO.getPackage_hotel_cetagory());
         }
+        PackageEntity packageEntity=byId.get();
+        packageEntity.setPackage_cetagory(packageDTO.getPackage_cetagory());
+        packageEntity.setPackage_vehicle_cetagory(packageDTO.getPackage_vehicle_cetagory());
+        packageEntity.setPackage_hotel_cetagory(packageDTO.getPackage_hotel_cetagory());
+         packageRepo.save(packageEntity);
 
     }
 
     @Override
     public void deletePackage(String package_id) {
- packageRepo.deleteById(package_id);
+        Optional<PackageEntity> byId = packageRepo.findById(package_id);
+        if (!byId.isPresent()){
+            throw new NotFoundException("Package ID Not Found :"+package_id);
+        }
+        packageRepo.deleteById(package_id);
     }
 }
