+++ src/main/java/org/example/service/util/TravelAreaServiceIMPL.java
@@ -2,6 +2,7 @@ package org.example.service.util;
 
 import org.example.dto.TravelAreaDTO;
 import org.example.entity.TravelAreaEntity;
+import org.example.exception.NotFoundException;
 import org.example.repo.TravelAreaRepo;
 import org.example.service.TravelAreaService;
 import org.example.util.Converter;
@@ -27,19 +28,31 @@ public class TravelAreaServiceIMPL implements TravelAreaService {
 
     @Override
     public TravelAreaDTO getSelectArea(String area_id) {
-        return converter.toAreaDto(travelAreaRepo.findById(area_id).get());
+        Optional<TravelAreaEntity> byId = travelAreaRepo.findById(area_id);
+        if (!byId.isPresent()){
+            throw new NotFoundException("Travel Area Id Is Not Found :"+area_id);
+        }
+        return converter.toAreaDto(byId.get());
     }
 
     @Override
-    public void updateArea(TravelAreaDTO areaDTO) {
-        Optional<TravelAreaEntity> byId = travelAreaRepo.findById(String.valueOf(areaDTO.getId()));
+    public void updateArea(String area_id,TravelAreaDTO areaDTO) {
+        Optional<TravelAreaEntity> byId = travelAreaRepo.findById(area_id);
         if (!byId.isPresent()){
-            byId.get().setTravel_area(areaDTO.getTravel_area());
+            throw new NotFoundException("Travel Area Id Is Not Found :"+area_id);
+//            byId.get().setTravel_area(areaDTO.getTravel_area());
         }
+        TravelAreaEntity travelArea=byId.get();
+        travelArea.setTravel_area(areaDTO.getTravel_area());
+        travelAreaRepo.save(travelArea);
     }
 
     @Override
     public void deleteArea(String area_id) {
-travelAreaRepo.deleteById(area_id);
+        Optional<TravelAreaEntity> byId = travelAreaRepo.findById(area_id);
+        if (!byId.isPresent()){
+            throw new NotFoundException("Travel Area Id Is Not Found :" + area_id);
+        }
+        travelAreaRepo.deleteById(area_id);
     }
 }
